Combine pages together for an opportunities page
