+++ src/App.tsx
@@ -2,12 +2,10 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 import "./style/App.css";
 
-import Opportunities from "./opportunities/pages/opportunities.tsx";
-
+import Opportunities from "./opportunities/pages/Opportunities.tsx";
 import Home from "./shared/pages/Home.tsx";
 import PageNotFound from "./shared/pages/404.tsx";
 import MainNavigation from "./shared/components/Navigation/MainNavigation.tsx";
-import Jobs from "./opportunities/pages/Jobs.tsx";
 import Departments from "./staff/pages/Departments.tsx";
 import StaffPage from "./staff/pages/Staff.tsx";
 import Department from "./staff/pages/Department.tsx";
@@ -37,10 +35,8 @@ function App() {
               <Route path="/login" element={<LoginRedirection />} />
               <Route path="/signout" element={<LogoutRedirection />} />
               <Route path="/logout" element={<LogoutRedirection />} />
-              
-              <Route path="/opportunities" element={<Opportunities />} />
 
-              <Route path="/jobs" element={<Jobs />} />
+              <Route path="/opportunities" element={<Opportunities />} />
               <Route path="/profile" element={<ProfilePage />} />
               <Route
                 path="/staff/department/:department"
